SALTO-3490: Restore ticket_form.ts to its original state
